refactor: rename bitcoin.ts as blockstream-explorer-links.ts
